feat: error handling for resource not found

+++ src/use-cases/subtasks/delete.ts
@@ -1,5 +1,6 @@
 import { SubtaskRepository } from "@/repositories/subtasks/subtask-repository"; 
 import { Subtask } from "@prisma/client";
+import { ResourceNotFoundError } from "../errors/resource-not-found-error";
 
 interface DeleteUseCaseRequest {
   id: string, 
@@ -8,9 +9,18 @@ interface DeleteUseCaseRequest {
 export class DeleteUseCase {
   constructor(private SubtaskRepository: SubtaskRepository) {}
 
+  
+
   async execute({
     id,    
   }:DeleteUseCaseRequest) : Promise<Subtask> {
+
+    const subtask = await this.SubtaskRepository.findById(id)
+
+    if(!subtask){
+      throw new ResourceNotFoundError()
+    }
+
     const deletedSubtask = await this.SubtaskRepository.delete(
       id
     )
