+++ src/use-cases/tasks/delete.ts
@@ -1,5 +1,6 @@
 import { TaskRepository } from "@/repositories/tasks/task-repository";
 import { Task } from "@prisma/client";
+import { ResourceNotFoundError } from "../errors/resource-not-found-error";
 
 interface DeleteUseCaseRequest {
   id: string, 
@@ -11,6 +12,13 @@ export class DeleteUseCase {
   async execute({
     id,    
   }:DeleteUseCaseRequest) : Promise<Task> {
+
+    const task = await this.TaskRepository.findById(id)
+
+    if(!task){
+      throw new ResourceNotFoundError()
+    }
+
     const deletedTask = await this.TaskRepository.delete(
       id
     )
